De-duplicate Istio Identification logic
This is needed to add pre-condition and eventually quota checks.
