+++ linkerd/protocol/h2/src/main/scala/io/buoyant/linkerd/protocol/h2/istio/H2IstioRequest.scala
@@ -4,7 +4,7 @@ import com.twitter.finagle.buoyant.h2.Request
 import io.buoyant.k8s.istio.IstioRequest
 
 object H2IstioRequest {
-  def apply(req: Request): IstioRequest =
-    new IstioRequest(req.path, req.scheme, req.method.toString, req.authority, req.headers.get)
+  def apply(req: Request): IstioRequest[Request] =
+    new IstioRequest(req.path, req.scheme, req.method.toString, req.authority, req.headers.get, req)
 }
