+++ linkerd/protocol/h2/src/main/scala/io/buoyant/linkerd/protocol/h2/istio/IstioIdentifier.scala
@@ -1,45 +1,32 @@
 package io.buoyant.linkerd.protocol.h2.istio
 
 import com.fasterxml.jackson.annotation.JsonIgnore
-import com.twitter.finagle.buoyant.Dst
 import com.twitter.finagle.buoyant.h2._
 import com.twitter.finagle.{Dtab, Path, Stack}
 import com.twitter.util.Future
 import io.buoyant.config.types.Port
 import io.buoyant.k8s.istio._
+import io.buoyant.k8s.istio.identifiers.InternalTrafficIdentifier
 import io.buoyant.k8s.istio.mixer.MixerClient
 import io.buoyant.linkerd.IdentifierInitializer
 import io.buoyant.linkerd.protocol.H2IdentifierConfig
-import io.buoyant.linkerd.protocol.h2.ErrorReseter.H2ResponseException
-import io.buoyant.router.RoutingFactory.{BaseDtab, DstPrefix, IdentifiedRequest, Identifier, RequestIdentification}
-import istio.proxy.v1.config.HTTPRedirect
+import io.buoyant.router.RoutingFactory.{BaseDtab, DstPrefix, Identifier, RequestIdentification}
 
-class IstioIdentifier(val pfx: Path, baseDtab: () => Dtab,
-  val routeCache: RouteCache,
-  val clusterCache: ClusterCache,
-  val mixerClient: MixerClient)
-  extends Identifier[Request] with IstioIdentifierBase[Request] {
+class IstioIdentifier(
+  pfx: Path,
+  baseDtab: () => Dtab,
+  routeCache: RouteCache,
+  clusterCache: ClusterCache,
+  mixerClient: MixerClient
+)
+  extends Identifier[Request] {
+  val istioIdentifier = new InternalTrafficIdentifier[Request](pfx, baseDtab, routeCache, clusterCache, mixerClient, new H2IstioRequestHandler)
 
-  override def apply(req: Request): Future[RequestIdentification[Request]] = {
-    getIdentifiedPath(req).map { path =>
-      val dst = Dst.Path(path, baseDtab(), Dtab.local)
-      new IdentifiedRequest(dst, req)
-    }
+  override def apply(originalRequest: Request): Future[RequestIdentification[Request]] = {
+    val req = H2IstioRequest(originalRequest)
+    istioIdentifier.identify(req)
   }
 
-  def redirectRequest(redir: HTTPRedirect, req: Request): Future[Nothing] = {
-    val resp = Response(Status.Found, Stream.empty())
-    resp.headers.set(Headers.Path, redir.`uri`.getOrElse(req.path))
-    resp.headers.set(Headers.Authority, redir.`authority`.getOrElse(req.authority))
-    Future.exception(H2ResponseException(resp))
-  }
-
-  def rewriteRequest(uri: String, authority: Option[String], req: Request): Unit = {
-    req.headers.set(Headers.Path, uri)
-    req.headers.set(Headers.Authority, authority.getOrElse(""))
-  }
-
-  def reqToMeta(req: Request): IstioRequest = H2IstioRequest(req)
 }
 
 case class IstioIdentifierConfig(
