+++ linkerd/protocol/h2/src/main/scala/io/buoyant/linkerd/protocol/h2/istio/IstioIngressIdentifier.scala
@@ -2,87 +2,36 @@ package io.buoyant.linkerd.protocol.h2.istio
 
 import com.fasterxml.jackson.annotation.JsonIgnore
 import com.twitter.finagle._
-import com.twitter.finagle.buoyant.Dst
-import com.twitter.finagle.buoyant.h2.{Headers, Request, Response, Status, Stream}
-import com.twitter.util.{Future, Try}
+import com.twitter.finagle.buoyant.h2.Request
+import com.twitter.util.Future
 import io.buoyant.config.types.Port
 import io.buoyant.k8s.IngressCache
-import io.buoyant.k8s.istio.ClusterCache.Cluster
+import io.buoyant.k8s.istio.identifiers.IngresssTrafficIdentifier
 import io.buoyant.k8s.istio.mixer.MixerClient
 import io.buoyant.k8s.istio.{IstioIngressConfigurator, _}
 import io.buoyant.linkerd.IdentifierInitializer
 import io.buoyant.linkerd.protocol.H2IdentifierConfig
-import io.buoyant.linkerd.protocol.h2.ErrorReseter.H2ResponseException
-import io.buoyant.router.RoutingFactory.{BaseDtab, DstPrefix, IdentifiedRequest, Identifier, RequestIdentification, UnidentifiedRequest}
-import istio.proxy.v1.config.HTTPRedirect
+import io.buoyant.router.RoutingFactory.{BaseDtab, DstPrefix, Identifier, RequestIdentification}
 
 class IstioIngressIdentifier(
-  val pfx: Path,
+  pfx: Path,
   baseDtab: () => Dtab,
   namespace: Option[String],
   apiClient: Service[http.Request, http.Response],
   annotationClass: String,
-  val routeCache: RouteCache,
-  val clusterCache: ClusterCache,
-  val mixerClient: MixerClient
-) extends Identifier[Request] with IstioIdentifierBase[Request] {
+  routeCache: RouteCache,
+  clusterCache: ClusterCache,
+  mixerClient: MixerClient
+) extends Identifier[Request] {
+
+  val istioIdentifier = new IngresssTrafficIdentifier[Request](pfx, baseDtab, routeCache, clusterCache, mixerClient, new H2IstioRequestHandler)
 
   private[this] val ingressCache = new IngressCache(namespace, apiClient, annotationClass)
 
   override def apply(req: Request): Future[RequestIdentification[Request]] = {
     val matchingPath = ingressCache.matchPath(Some(req.authority), req.path)
-    matchingPath.flatMap {
-      case None =>
-        Future.value(new UnidentifiedRequest(s"no ingress rule matches ${req.authority}:${req.path}"))
-      case Some(ingressPath) =>
-        val clusterName = s"${ingressPath.svc}.${ingressPath.namespace}.svc.cluster.local"
-
-        // use clusterCache to transform any port Int into a port name
-        val portName = Try(ingressPath.port.toInt).toOption match {
-          case Some(portNumber) => clusterCache.get(s"$clusterName:$portNumber").map {
-            case Some(Cluster(_, p)) => Some(p)
-            case None => None
-          }
-          case None => Future.value(Some(ingressPath.port))
-        }
-
-        Future.join(portName, routeCache.getRules).flatMap {
-          case (Some(port), rules) =>
-            val filteredRules = filterRules(rules, clusterName, reqToMeta(req))
-            (maxPrecedenceRule(filteredRules) match {
-              case Some((ruleName, rule)) =>
-                rule.`redirect` match {
-                  case Some(redir) => redirectRequest(redir, req)
-                  case None =>
-                    val (uri, authority) = httpRewrite(rule, reqToMeta(req))
-                    rewriteRequest(uri, authority, req)
-                    Future.value(pfx ++ Path.Utf8("route", ruleName, port))
-                }
-              //forward requests which have no matching rules to an empty label selector
-              case None => Future.value(pfx ++ Path.Utf8("dest", clusterName, "::", port))
-            }).map { path =>
-              val dst = Dst.Path(path, baseDtab(), Dtab.local)
-              new IdentifiedRequest(dst, req)
-            }
-          case _ => Future.value(new UnidentifiedRequest(s"ingress path ${ingressPath.svc}:${ingressPath.port} does not match any istio vhosts"))
-        }
-    }
+    istioIdentifier.identify(H2IstioRequest(req), matchingPath)
   }
-
-  def redirectRequest(redir: HTTPRedirect, req: Request): Future[Nothing] = {
-    val resp = Response(Status.Found, Stream.empty())
-    resp.headers.set(Headers.Path, redir.`uri`.getOrElse(req.path))
-    resp.headers.set(Headers.Authority, redir.`authority`.getOrElse(req.authority))
-    Future.exception(H2ResponseException(resp))
-  }
-
-  def rewriteRequest(uri: String, authority: Option[String], req: Request): Unit = {
-    req.headers.set(Headers.Path, uri)
-    req.headers.set(Headers.Authority, authority.getOrElse(""))
-  }
-
-  def reqToMeta(req: Request): IstioRequest = H2IstioRequest(req)
-
 }
 
 case class IstioIngressIdentifierConfig(
