+++ linkerd/protocol/h2/src/test/scala/io/buoyant/linkerd/protocol/h2/istio/H2IstioResponseTest.scala
@@ -1,10 +1,9 @@
 package io.buoyant.linkerd.protocol.h2.istio
 
-import com.twitter.finagle.buoyant.h2.{Response, Status}
+import com.twitter.finagle.buoyant.h2.{Response, Status, Stream}
 import com.twitter.util.{Duration, Return, Throw}
 import io.buoyant.k8s.istio.{ResponseDurationIstioAttribute, SourceLabelIstioAttribute, TargetLabelsIstioAttribute, TargetServiceIstioAttribute}
 import org.scalatest.FunSuite
-import com.twitter.finagle.buoyant.h2.Stream
 
 class H2IstioResponseTest extends FunSuite {
   {
@@ -23,7 +22,6 @@ class H2IstioResponseTest extends FunSuite {
     }
 
     test("generates istio response for failed response") {
-      val httpResponse = Response(Status.Ok, Stream())
       val duration = Duration.Top
       val response = Throw(new UnsupportedOperationException("not implemented"))
       val istioResponse = H2IstioResponse(response, duration)
