+++ linkerd/protocol/h2/src/test/scala/io/buoyant/linkerd/protocol/h2/istio/IstioLoggerTest.scala
@@ -4,11 +4,10 @@ import com.twitter.finagle.buoyant.H2
 import com.twitter.finagle.buoyant.h2.{Headers, Request, Response, Status, Stream => H2Stream}
 import com.twitter.finagle.{Service, Stack}
 import com.twitter.util.Future
-import io.buoyant.grpc.runtime.Stream
 import io.buoyant.k8s.istio._
 import io.buoyant.k8s.istio.mixer.MixerClient
 import io.buoyant.test.Awaits
-import istio.mixer.v1.{Mixer, ReportResponse}
+import istio.mixer.v1.Mixer
 import org.scalatest.FunSuite
 
 class IstioLoggerTest extends FunSuite with Awaits {
