+++ linkerd/protocol/http/src/main/scala/io/buoyant/linkerd/protocol/http/istio/HttpIstioRequest.scala
@@ -4,7 +4,8 @@ import com.twitter.finagle.http.Request
 import io.buoyant.k8s.istio.IstioRequest
 
 object HttpIstioRequest {
-  def apply(req: Request): IstioRequest =
-    IstioRequest(req.path, "", req.method.toString, req.host.getOrElse(""), req.headerMap.get)
+  def apply(req: Request): IstioRequest[Request] =
+    //TODO: match on request scheme
+    IstioRequest(req.path, "", req.method.toString, req.host.getOrElse(""), req.headerMap.get, req)
 }
