+++ linkerd/protocol/http/src/main/scala/io/buoyant/linkerd/protocol/http/istio/IstioIngressIdentifier.scala
@@ -2,87 +2,34 @@ package io.buoyant.linkerd.protocol.http.istio
 
 import com.fasterxml.jackson.annotation.JsonIgnore
 import com.twitter.finagle._
-import com.twitter.finagle.buoyant.Dst
-import com.twitter.finagle.http.{Request, Response, Status}
-import com.twitter.util.{Future, Try}
+import com.twitter.finagle.http.Request
+import com.twitter.util.Future
 import io.buoyant.config.types.Port
 import io.buoyant.k8s.IngressCache
-import io.buoyant.k8s.istio.ClusterCache.Cluster
 import io.buoyant.k8s.istio._
+import io.buoyant.k8s.istio.identifiers.{IngresssTrafficIdentifier, IstioProtocolSpecificRequestHandler}
 import io.buoyant.k8s.istio.mixer.MixerClient
 import io.buoyant.linkerd.IdentifierInitializer
 import io.buoyant.linkerd.protocol.HttpIdentifierConfig
-import io.buoyant.linkerd.protocol.http.ErrorResponder.HttpResponseException
-import io.buoyant.router.RoutingFactory.{IdentifiedRequest, Identifier, RequestIdentification, UnidentifiedRequest}
-import istio.proxy.v1.config.HTTPRedirect
+import io.buoyant.router.RoutingFactory.{Identifier, RequestIdentification}
 
 class IstioIngressIdentifier(
-  val pfx: Path,
+  pfx: Path,
   baseDtab: () => Dtab,
   namespace: Option[String],
   apiClient: Service[http.Request, http.Response],
   annotationClass: String,
-  val routeCache: RouteCache,
-  val clusterCache: ClusterCache,
-  val mixerClient: MixerClient
-) extends Identifier[Request] with IstioIdentifierBase[Request] {
-
+  routeCache: RouteCache,
+  clusterCache: ClusterCache,
+  mixerClient: MixerClient,
+  requestHandler: IstioProtocolSpecificRequestHandler[Request]
+) extends Identifier[Request] {
+  val internalIstioIdentifier = new IngresssTrafficIdentifier[Request](pfx, baseDtab, routeCache, clusterCache, mixerClient, new HttpIstioRequestHandler)
   private[this] val ingressCache = new IngressCache(namespace, apiClient, annotationClass)
 
   override def apply(req: Request): Future[RequestIdentification[Request]] = {
     val matchingPath = ingressCache.matchPath(req.host, req.path)
-    matchingPath.flatMap {
-      case None =>
-        Future.value(new UnidentifiedRequest(s"no ingress rule matches ${req.host}:${req.path}"))
-      case Some(ingressPath) =>
-        val clusterName = s"${ingressPath.svc}.${ingressPath.namespace}.svc.cluster.local"
-
-        // use clusterCache to transform any port Int into a port name
-        val portName = Try(ingressPath.port.toInt).toOption match {
-          case Some(portNumber) => clusterCache.get(s"$clusterName:$portNumber").map {
-            case Some(Cluster(_, p)) => Some(p)
-            case None => None
-          }
-          case None => Future.value(Some(ingressPath.port))
-        }
-
-        Future.join(portName, routeCache.getRules).flatMap {
-          case (Some(port), rules) =>
-            val filteredRules = filterRules(rules, clusterName, reqToMeta(req))
-            (maxPrecedenceRule(filteredRules) match {
-              case Some((ruleName, rule)) =>
-                rule.`redirect` match {
-                  case Some(redir) => redirectRequest(redir, req)
-                  case None =>
-                    val (uri, authority) = httpRewrite(rule, reqToMeta(req))
-                    rewriteRequest(uri, authority, req)
-                    Future.value(pfx ++ Path.Utf8("route", ruleName, port))
-                }
-              //forward requests which have no matching rules to an empty label selector
-              case None => Future.value(pfx ++ Path.Utf8("dest", clusterName, "::", port))
-            }).map { path =>
-              val dst = Dst.Path(path, baseDtab(), Dtab.local)
-              new IdentifiedRequest(dst, req)
-            }
-          case _ => Future.value(new UnidentifiedRequest(s"ingress path ${ingressPath.svc}:${ingressPath.port} does not match any istio vhosts"))
-        }
-    }
-  }
-
-  def reqToMeta(req: Request): IstioRequest =
-    //TODO: match on request scheme
-    IstioRequest(req.path, "", req.method.toString, req.host.getOrElse(""), req.headerMap.get)
-
-  def redirectRequest(redir: HTTPRedirect, req: Request): Future[Nothing] = {
-    val redirect = Response(Status.Found)
-    redirect.location = redir.`uri`.getOrElse(req.uri)
-    redirect.host = redir.`authority`.orElse(req.host).getOrElse("")
-    Future.exception(HttpResponseException(redirect))
-  }
-
-  def rewriteRequest(uri: String, authority: Option[String], req: Request): Unit = {
-    req.uri = uri
-    req.host = authority.getOrElse("")
+    internalIstioIdentifier.identify(HttpIstioRequest(req), matchingPath)
   }
 
 }
@@ -116,7 +63,8 @@ case class IstioIngressIdentifierConfig(
       IngressAnnotationClass,
       mkRouteCache(host, port),
       mkClusterCache(discoveryHost, discoveryPort),
-      mkMixerClient(mixerHost, mixerPort)
+      mkMixerClient(mixerHost, mixerPort),
+      new HttpIstioRequestHandler
     )
   }
