+++ linkerd/protocol/http/src/test/scala/io/buoyant/linkerd/protocol/http/istio/IstioIdentifierTest.scala
@@ -6,8 +6,8 @@ import com.twitter.finagle.http.{Request => FRequest, Response => FResponse}
 import com.twitter.finagle.{Dtab, Path, Service}
 import com.twitter.util.Future
 import io.buoyant.grpc.runtime.GrpcStatus
-import io.buoyant.k8s.istio.mixer.{MixerCheckStatus, MixerClient}
 import io.buoyant.k8s.istio._
+import io.buoyant.k8s.istio.mixer.{MixerCheckStatus, MixerClient}
 import io.buoyant.router.RoutingFactory._
 import io.buoyant.test.Awaits
 import org.scalatest.FunSuite
@@ -190,7 +190,7 @@ class IstioIdentifierTest extends FunSuite with Awaits {
       duration: ResponseDurationIstioAttribute
     ) = Future.Done
 
-    override def checkPreconditions(istioRequest: IstioRequest) = Future.value(MixerCheckStatus(GrpcStatus.Ok("")))
+    override def checkPreconditions(istioRequest: IstioRequest[_]) = Future.value(MixerCheckStatus(GrpcStatus.Ok("")))
   }
 
   val identifier = new IstioIdentifier(Path.Utf8("svc"), () => Dtab.base, routeCache, clusterCache, noOpMixerClient)
